Extract class InMemoryTransactionLogSnapshotSetup
